# Replace boxing ArrayList in PathParser, improve performance, decrease allocs