--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/Data/PathParser.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/Data/PathParser.cs
@@ -6,9 +6,11 @@
 // Description: Parser for the Path of a (CLR) binding
 //
 
-using System.Collections;
+using System;
 using System.Text;          // StringBuilder
 using System.Windows;
+using System.Collections.Generic;
+
 using MS.Utility;           // FrugalList
 
 namespace MS.Internal.Data
@@ -58,11 +60,11 @@ namespace MS.Internal.Data
 
     internal class PathParser
     {
-        string _error;
-        public String Error { get { return _error; } }
-        void SetError(string id, params object[] args) { _error = SR.Format(SR.GetResourceString(id), args); }
+        private string _error;
+        public string Error { get { return _error; } }
+        private void SetError(string id, params object[] args) { _error = SR.Format(SR.GetResourceString(id), args); }
 
-        enum State { Init, DrillIn, Prop, Done };
+        private enum State { Init, DrillIn, Prop, Done };
 
         // Each level of the path consists of
         //      a property or indexer:
@@ -80,7 +82,7 @@ namespace MS.Internal.Data
 
         public SourceValueInfo[] Parse(string path)
         {
-            _path = (path != null) ? path.Trim() : String.Empty;
+            _path = (path != null) ? path.Trim() : string.Empty;
             _n = _path.Length;
 
             if (_n == 0)
@@ -93,14 +95,14 @@ namespace MS.Internal.Data
             _index = 0;
             _drillIn = DrillIn.IfNeeded;
 
-            _al.Clear();
+            _sourceValueInfos.Clear();
             _error = null;
             _state = State.Init;
 
             while (_state != State.Done)
             {
                 char c = (_index < _n) ? _path[_index] : NullChar;
-                if (Char.IsWhiteSpace(c))
+                if (char.IsWhiteSpace(c))
                 {
                     ++_index;
                     continue;
@@ -138,7 +140,7 @@ namespace MS.Internal.Data
                                 break;
                             default:
                                 SetError(nameof(SR.PathSyntax), _path.Substring(0, _index), _path.Substring(_index));
-                                return EmptyInfo;
+                                return s_emptyInfo;
                         }
                         _state = State.Prop;
                         break;
@@ -168,12 +170,12 @@ namespace MS.Internal.Data
 
             if (_error == null)
             {
-                result = new SourceValueInfo[_al.Count];
-                _al.CopyTo(result);
+                result = new SourceValueInfo[_sourceValueInfos.Count];
+                _sourceValueInfos.CopyTo(result);
             }
             else
             {
-                result = EmptyInfo;
+                result = s_emptyInfo;
             }
 
             return result;
@@ -216,15 +218,15 @@ namespace MS.Internal.Data
                 ? new SourceValueInfo(SourceValueType.Property, _drillIn, name)
                 : new SourceValueInfo(SourceValueType.Direct, _drillIn, (string)null);
 
-            _al.Add(info);
+            _sourceValueInfos.Add(info);
 
             StartNewLevel();
         }
 
 
-        enum IndexerState { BeginParam, ParenString, ValueString, Done }
+        private enum IndexerState { BeginParam, ParenString, ValueString, Done }
 
-        void AddIndexer()
+        private void AddIndexer()
         {
             // indexer args are parsed by a (sub-) state machine with four
             // states.  The string is a comma-separated list of params, each
@@ -253,7 +255,7 @@ namespace MS.Internal.Data
                     return;
                 }
 
-                Char c = _path[_index++];
+                char c = _path[_index++];
 
                 // handle the escape character - set the flag for the next character
                 if (c == EscapeChar && !escaped)
@@ -276,7 +278,7 @@ namespace MS.Internal.Data
                             // '(' introduces optional paren string
                             state = IndexerState.ParenString;
                         }
-                        else if (Char.IsWhiteSpace(c))
+                        else if (char.IsWhiteSpace(c))
                         {
                             // ignore leading white space
                         }
@@ -324,7 +326,7 @@ namespace MS.Internal.Data
                                 --level;
                             }
                         }
-                        else if (Char.IsWhiteSpace(c))
+                        else if (char.IsWhiteSpace(c))
                         {
                             // add white space, but trim it later if it's trailing
                             valueStringBuilder.Append(c);
@@ -371,10 +373,8 @@ namespace MS.Internal.Data
             }
 
             // assemble the final result
-            SourceValueInfo info = new SourceValueInfo(
-                                        SourceValueType.Indexer,
-                                        _drillIn, paramList);
-            _al.Add(info);
+            SourceValueInfo info = new(SourceValueType.Indexer, _drillIn, paramList);
+            _sourceValueInfos.Add(info);
 
             StartNewLevel();
         }
@@ -390,15 +390,18 @@ namespace MS.Internal.Data
             return ch == '.' || ch == '/' || ch == '[' || ch == ']';
         }
 
-        State _state;
-        string _path;
-        int _index;
-        int _n;
-        DrillIn _drillIn;
-        ArrayList _al = new ArrayList();
-        const char NullChar = Char.MinValue;
-        const char EscapeChar = '^';
-        static SourceValueInfo[] EmptyInfo = Array.Empty<SourceValueInfo>();
+        private State _state;
+        private string _path;
+        private int _index;
+        private int _n;
+        private DrillIn _drillIn;
+
+        private const char NullChar = char.MinValue;
+        private const char EscapeChar = '^';
+
+        private readonly List<SourceValueInfo> _sourceValueInfos = new();
+
+        private static readonly SourceValueInfo[] s_emptyInfo = Array.Empty<SourceValueInfo>();
     }
 }