--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/Data/PathParser.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/Data/PathParser.cs
@@ -147,20 +147,15 @@ namespace MS.Internal.Data
                         break;
 
                     case State.Prop:
-                        bool isIndexer = false;
                         switch (c)
                         {
-                            case '[':
-                                isIndexer = true;
+                            case '[': // Indexer follows
+                                AddIndexer();
                                 break;
-                            default:
+                            default: // Property follows
+                                AddProperty();
                                 break;
-                        }
-
-                        if (isIndexer)
-                            AddIndexer();
-                        else
-                            AddProperty();
+                        }                           
 
                         break;
                 }