## Description

Replaces boxing `ArrayList` with `List<SourceValueInfo>`, greatly improving performance and decreasing allocations on this kind of hot path. Also removes static field of composing `Array.Empty<SourceValueInfo>()` as it has no impact on storing it.

I've also removed the `bool isIndexer` as it just complicates the `switch` statement, added explicit accessibility modifiers, and sealed the class as it has no derivates.

### 10 items added and copied

| Method   | error | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|--------- |------ |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original | True  | 139.60 ns |   2.808 ns |    3.748 ns | 0.0491 |       2,314 B |         824 B |
| PR__EDIT | True  |  42.81 ns |   0.846 ns |    0.831 ns | 0.0206 |       2,155 B |         344 B |

### 2 items added and copied

| Method   | error | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|--------- |------ |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original | True  |  45.55 ns |   0.872 ns |    0.933 ns | 0.0110 |       1,113 B |         184 B |
| PR__EDIT | True  |  16.34 ns |   0.285 ns |    0.266 ns | 0.0052 |       1,094 B |          88 B |

### 1 item added and copied

| Method   | error | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|--------- |------ |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original | True  |  33.50 ns |   0.646 ns |    0.604 ns | 0.0062 |         953 B |         104 B |
| PR__EDIT | True  |  13.36 ns |   0.176 ns |    0.165 ns | 0.0033 |         974 B |          56 B |

### State machine-like scenario

| Method   | error | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|--------- |------ |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original | True  |  385.3 ns |    4.13 ns |     3.87 ns | 0.0620 |         980 B |        1040 B |
| PR__EDIT | True  |  135.7 ns |    1.11 ns |     0.92 ns | 0.0334 |       1,006 B |         560 B |

### Benchmark code

<details><summary>Benchmark code</summary>

```csharp
[GlobalSetup]
public  void Setup()
{
    arrayList = new(10);
    genericList = new(10);
}

private ArrayList arrayList;
private List<SourceValueInfo> genericList;

//Loop like-scenario is like:
//SourceValueInfo[] result = null;
//for (int i = 0; i < 10; i++)
//{
//list.Clear(); 
//etc. etc.
//}
//return result;

[Benchmark]
[Arguments(true)]
public SourceValueInfo[] Original(bool error)
{
    arrayList.Clear();

    SourceValueInfo[] result;

    arrayList.Add(new SourceValueInfo(0, 1, (string)null));
    arrayList.Add(new SourceValueInfo(0, 1, (string)null));
    arrayList.Add(new SourceValueInfo(0, 1, (string)null));
    arrayList.Add(new SourceValueInfo(0, 1, (string)null));
    arrayList.Add(new SourceValueInfo(0, 1, (string)null));
    arrayList.Add(new SourceValueInfo(0, 1, (string)null));
    arrayList.Add(new SourceValueInfo(0, 1, (string)null));
    arrayList.Add(new SourceValueInfo(0, 1, (string)null));
    arrayList.Add(new SourceValueInfo(0, 1, (string)null));
    arrayList.Add(new SourceValueInfo(0, 1, (string)null));

    if (error)
    {
        result = new SourceValueInfo[arrayList.Count];
        arrayList.CopyTo(result);
    }
    else
    {
        result = Array.Empty<SourceValueInfo>();
    }

    return result;
}

[Benchmark]
[Arguments(true)]
public SourceValueInfo[] PR__EDIT(bool error)
{
    genericList.Clear();

    SourceValueInfo[] result;

    genericList.Add(new SourceValueInfo(0, 1, (string)null));
    genericList.Add(new SourceValueInfo(0, 1, (string)null));
    genericList.Add(new SourceValueInfo(0, 1, (string)null));
    genericList.Add(new SourceValueInfo(0, 1, (string)null));
    genericList.Add(new SourceValueInfo(0, 1, (string)null));
    genericList.Add(new SourceValueInfo(0, 1, (string)null));
    genericList.Add(new SourceValueInfo(0, 1, (string)null));
    genericList.Add(new SourceValueInfo(0, 1, (string)null));
    genericList.Add(new SourceValueInfo(0, 1, (string)null));
    genericList.Add(new SourceValueInfo(0, 1, (string)null));       

    if (error)
    {
        result = new SourceValueInfo[genericList.Count];
        genericList.CopyTo(result);
    }
    else
    {
        result = Array.Empty<SourceValueInfo>();
    }

    return result;
}
```

</details>

## Customer Impact

Greatly improved performance, nicely decreased allocations in `DataBindingEngine`.

## Regression

No.

## Testing

Local build.

## Risk

Low, just a collection type swap.

 ###### Microsoft Reviewers: [Open in CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/dotnet/wpf/pull/9631)